# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER ULBI",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/jul003",
            "email": "indra@ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "ws-jul2024-bdc7cc35640a.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/deletegadget/{id}": {
            "delete": {
                "description": "Hapus data gadget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gadget"
                ],
                "summary": "Delete data Gadget.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/gadget": {
            "get": {
                "description": "Mengambil semua data gadget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gadget"
                ],
                "summary": "Get All Data Gadget.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Gadget"
                        }
                    }
                }
            }
        },
        "/insertgadget": {
            "post": {
                "description": "Input data gadget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gadget"
                ],
                "summary": "Insert data Gadget.",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqGadget"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Gadget"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/insertreview": {
            "post": {
                "description": "Input data review.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "Insert data Review.",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Review"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/review": {
            "get": {
                "description": "Mengambil semua data review.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "Get All Data Review.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Gadget"
                        }
                    }
                }
            }
        },
        "/updategadget/{id}": {
            "put": {
                "description": "Ubah data gadget.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gadget"
                ],
                "summary": "Update data Gadegt.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqGadget"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Gadget"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Gadget": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "deskripsi": {
                    "type": "string"
                },
                "harga": {
                    "type": "number"
                },
                "merk": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "spesifikasi": {
                    "$ref": "#/definitions/controller.Spesifikasi"
                }
            }
        },
        "controller.ReqGadget": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "deskripsi": {
                    "type": "string"
                },
                "harga": {
                    "type": "number"
                },
                "merk": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "spesifikasi": {
                    "$ref": "#/definitions/controller.Spesifikasi"
                }
            }
        },
        "controller.ReqReview": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "gadget_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "review": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "controller.Review": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "datetime": {
                    "description": "Use time.Time instead of primitive.DateTime",
                    "type": "string"
                },
                "gadget_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "review": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "controller.Spesifikasi": {
            "type": "object",
            "properties": {
                "baterai": {
                    "type": "string"
                },
                "fitur_lainnya": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kamera": {
                    "type": "string"
                },
                "layar": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "prosesor": {
                    "type": "string"
                },
                "ram": {
                    "type": "integer"
                },
                "storage": {
                    "type": "integer"
                }
            }
        }
    }
}